Refactor | #168 | @YongsHub | 의존 방향 변경
