+++ cakk-external/src/main/java/com/cakk/external/executor/CertificationApiExecutor.java
@@ -1,4 +1,4 @@
-package com.cakk.external.api;
+package com.cakk.external.executor;
 
 public interface CertificationApiExecutor<T> {
 	void send(T message);
