+++ cakk-external/src/main/java/com/cakk/external/extractor/CertificationSlackMessageExtractor.java
@@ -1,4 +1,4 @@
-package com.cakk.external.api;
+package com.cakk.external.extractor;
 
 import java.util.List;
